refactor to clean up petal shapes a bit, prep for manual pointy and prism petals
