+++ webpack/containers/shapes/RoundedPetal.js
@@ -8,11 +8,8 @@ class RoundedPetal extends AbstractPetal {
         super(props);
 
     }
-    drawPetal(startInnerPoint, startOuterPoint, startControlPoint,
-        returnInnerPoint, returnOuterPoint, returnControlPoint) {
-
-        return drawRoundedPetal(startInnerPoint, startOuterPoint, startControlPoint,
-            returnInnerPoint, returnOuterPoint, returnControlPoint);
+    drawPetal() {
+        return drawRoundedPetal(...arguments);
     }
     render() {
         const paths = this.singleControlPointMirroredForEachPetalArmAlgorithm();
