+++ webpack/petalUtil.js
@@ -75,6 +75,34 @@ export const drawCurveyPetal = (startInnerPoint, startOuterPoint, startInnerCont
     return path.str();
 };
 
+export const drawPointedPetal = (startInnerPoint, startOuterPoint, startControlPoint,
+    returnInnerPoint, returnOuterPoint, returnControlPoint) => {
+    const path = SvgPath().to(...startInnerPoint[0])
+        .line(...startControlPoint[0])
+        .line(...startOuterPoint[0])
+        .line(...returnOuterPoint[0])
+        .line(...returnControlPoint[0])
+        .line(...returnInnerPoint[0])
+        .close();
+
+    return path.str();
+};
+
+export const drawPrismPetal = (startInnerPoint, startOuterPoint, startInnerControlPoint, startOuterControlPoint,
+    returnInnerPoint, returnOuterPoint, returnInnerControlPoint, returnOuterControlPoint) => {
+    const path = SvgPath().to(...startInnerPoint[0])
+        .line(...startInnerControlPoint[0])
+        .line(...startOuterControlPoint[0])
+        .line(...startOuterPoint[0])
+        .line(...returnOuterPoint[0])
+        .line(...returnOuterControlPoint[0])
+        .line(...returnInnerControlPoint[0])
+        .line(...returnInnerPoint[0])
+        .close();
+
+    return path.str();
+}
+
 const calculatePetalPoints = (drawPetal, points, angle, centerPoint, rightPoints) => {
     const pointValues = points.map(point => [Object.values(point)]);
     const twirls = pointValues.map(values => twirl.rotateZoom(angle, centerPoint, 1, values));
